Moves `URLProtocolStub` to separate file.
